[vm, compiler] Remove surprising truncation from x64's LoadImmediate.
All other architectures have a LoadImmediate that works with any uword, but x64's truncated to 32-bit, causing surprising x64-only failures when LoadImmediate is used from architecture-independent parts of the compiler.

TEST=ci
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/268040
Commit-Queue: Ryan Macnak <rmacnak@google.com>

+++ runtime/vm/compiler/assembler/assembler_x64.h
@@ -566,7 +566,7 @@ class Assembler : public AssemblerBase {
                         const Immediate& imm,
                         OperandSize width = kEightBytes);
   void CompareImmediate(Register reg,
-                        int32_t immediate,
+                        int64_t immediate,
                         OperandSize width = kEightBytes) {
     return CompareImmediate(reg, Immediate(immediate), width);
   }
@@ -583,11 +583,11 @@ class Assembler : public AssemblerBase {
                      OperandSize width = kEightBytes);
 
   void AndImmediate(Register dst, const Immediate& imm);
-  void AndImmediate(Register dst, int32_t value) {
+  void AndImmediate(Register dst, int64_t value) {
     AndImmediate(dst, Immediate(value));
   }
   void OrImmediate(Register dst, const Immediate& imm);
-  void OrImmediate(Register dst, int32_t value) {
+  void OrImmediate(Register dst, int64_t value) {
     OrImmediate(dst, Immediate(value));
   }
   void XorImmediate(Register dst, const Immediate& imm);
@@ -767,7 +767,7 @@ class Assembler : public AssemblerBase {
                     const Immediate& imm,
                     OperandSize width = kEightBytes);
   void AddImmediate(Register reg,
-                    int32_t value,
+                    int64_t value,
                     OperandSize width = kEightBytes) {
     AddImmediate(reg, Immediate(value), width);
   }
@@ -780,7 +780,7 @@ class Assembler : public AssemblerBase {
                  int32_t value) {
     leaq(dest, Address(src, scale, value));
   }
-  void AddImmediate(Register dest, Register src, int32_t value);
+  void AddImmediate(Register dest, Register src, int64_t value);
   void AddImmediate(const Address& address, const Immediate& imm);
   void SubImmediate(Register reg,
                     const Immediate& imm,
@@ -797,7 +797,7 @@ class Assembler : public AssemblerBase {
 
   // Unlike movq this can affect the flags or use the constant pool.
   void LoadImmediate(Register reg, const Immediate& imm);
-  void LoadImmediate(Register reg, int32_t immediate) {
+  void LoadImmediate(Register reg, int64_t immediate) {
     LoadImmediate(reg, Immediate(immediate));
   }
   void LoadDImmediate(FpuRegister dst, double immediate);
